Bug fix: change return type of GetPreviousItemIndex to int from size_t.
NavigationManagerImpl::GetPreviousItemIndex is a wrapper around
|CRWSessionController previousItemIndex|. This can actually be -1, so
the return type should be int not size_t. This mistake was introduced
in https://codereview.chromium.org/2944093002.

Added a unit test to verify this edge case. Moved the method to public
section to allow access in test.

Bug: 734150

+++ ios/web/navigation/legacy_navigation_manager_impl.mm
@@ -456,7 +456,7 @@ LegacyNavigationManagerImpl::GetLastCommittedNonAppSpecificItem() const {
   return nullptr;
 }
 
-size_t LegacyNavigationManagerImpl::GetPreviousItemIndex() const {
+int LegacyNavigationManagerImpl::GetPreviousItemIndex() const {
   return [session_controller_ previousItemIndex];
 }
